requirements.txt
mistralai
smolagents
python-dotenv
